+++ NEWS.md
@@ -1,3 +1,10 @@
+# findSVI 0.2.0
+
+* findSVI now supports SVI calculation for 2022 (#20).
+
+* A new function `find_svi_x()` supports SVI calculation at custom geographic boundaries, with a user-defined crosswalk to Census geographies. Also available through `get_census_data()` with an additional argument `exp=TRUE` and `get_svi_x()` (#16).
+
+
 # findSVI 0.1.2
 
 * Wrapping data retrieval and SVI calculation, `find_svi()` supports multiple entries of year-state pairs.